feat: ajuste no cálculo das férias indenizadas e mudança do cálculo da rescisão para a pasta hooks, para correta separação
