+++ src/pages/Layout/styles.css
@@ -3,7 +3,7 @@
     display: flex;
     min-height: 100vh;
     flex-direction: column;
-    overflow-y: hidden;
+    overflow-x: hidden;
 }
 
 .content {
